attached_assets/requirements_1754645800475.txt
flask
flask-cors
openai
scikit-learn
pandas
joblib
psycopg2-binary
